M4 Wave 2 B3: strict flags for value, apt/apt, textaux, supportviews, support, lookz, textobjects, extensions, frame; fixes 11 real LP64/signal-handler/writable-string bugs — closes Wave 2
FossilOrigin-Name: 2f65c929c0953305463928637cde7fd7a630129f4df0a0e14bcdea644da11173

+++ src/atk/value/value.c
@@ -186,7 +186,7 @@ long value__Write(struct value *self, FILE *file, long writeID, int level)
     if (value_GetWriteID(self) != writeID)  {
 	value_SetWriteID( self,writeID);
         fprintf(file, "\\begindata{%s,%ld}\n", class_GetTypeName(self),value_GetID(self));
-	fprintf(file,">%d\n",value_GetValue(self));
+	fprintf(file,">%ld\n",value_GetValue(self));
         fprintf(file, "\\enddata{%s,%ld}\n",  class_GetTypeName(self),value_GetID(self));
     }
 
@@ -198,7 +198,7 @@ long value__Read(struct value *self, FILE *file, long id)
     while(1){
 	if(fgets(buf,256,file) == NULL) return dataobject_PREMATUREEOF;
 	if(*buf == '>') {
-	    self->rock1 = atoi(buf + 1);
+	    self->rock1 = atol(buf + 1);
 	    value_Put(self,valueatom,rock1atom,self->rock1);
 	}
 	else if(strncmp(buf,"\\enddata",8) == 0)return dataobject_NOREADERROR;
